Created a mock service with apiSauce and moved the service related interfaces/DTO
