+++ .env
@@ -1,0 +1,2 @@
+VITE_API_URL=http://localhost:3000
+VITE_API_KEY=supersecretkey
