+++ package.json
@@ -14,6 +14,7 @@
     "@emotion/styled": "^11.8.1",
     "@mantine/core": "^4.1.1",
     "@mantine/hooks": "^4.1.1",
+    "apisauce": "^2.1.5",
     "react": "^18.0.0",
     "react-dom": "^18.0.0"
   },
