+++ src/services/index.ts
@@ -1,0 +1,3 @@
+import { getCompaniesService } from './companies/companies';
+
+export { getCompaniesService };
